Implement Validating, Decrypting, and Encrypting MION firmware files (#17)
* wip: decrypting/re-encrypting firmware

* upgrade to hyper 1.0 (aka reqwest), clippy

* update windows handle type changes with new versions

* implement mion firmware parsing

this implements the actual base interfaces for managing MION FW
files. We can now decrypt, re-encrypt, and properly checksum/validate
any firmware images that would have been uploaded to a MION device.

also fixed some formatting nits

* fix last broken isize cast

* fix up some error message documentation

+++ pkg/cat-dev/src/serial/underlying/sys/windows.rs
@@ -77,7 +77,7 @@ impl RawSyncSerialPort {
 			WriteTotalTimeoutConstant: DEFAULT_TIMEOUT_MS,
 		};
 		unsafe {
-			SetCommTimeouts(HANDLE(fd.as_raw_handle() as isize), &timeouts)
+			SetCommTimeouts(HANDLE(fd.as_raw_handle()), &timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		let dcb = DCB {
@@ -89,8 +89,7 @@ impl RawSyncSerialPort {
 			..Default::default()
 		};
 		unsafe {
-			SetCommState(HANDLE(fd.as_raw_handle() as isize), &dcb)
-				.map_err(|_| IoError::last_os_error())?;
+			SetCommState(HANDLE(fd.as_raw_handle()), &dcb).map_err(|_| IoError::last_os_error())?;
 		}
 
 		Ok(Self { fd })
@@ -115,7 +114,7 @@ impl RawSyncSerialPort {
 	pub fn get_read_timeout(&self) -> IoResult<Duration> {
 		let mut timeouts = unsafe { std::mem::zeroed() };
 		unsafe {
-			GetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &mut timeouts)
+			GetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &mut timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		Ok(Duration::from_millis(
@@ -131,7 +130,7 @@ impl RawSyncSerialPort {
 	pub fn get_write_timeout(&self) -> IoResult<Duration> {
 		let mut timeouts = unsafe { std::mem::zeroed() };
 		unsafe {
-			GetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &mut timeouts)
+			GetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &mut timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		Ok(Duration::from_millis(
@@ -156,14 +155,14 @@ impl RawSyncSerialPort {
 
 		let mut timeouts = unsafe { std::mem::zeroed() };
 		unsafe {
-			GetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &mut timeouts)
+			GetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &mut timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		timeouts.ReadIntervalTimeout = u32::MAX;
 		timeouts.ReadTotalTimeoutMultiplier = u32::MAX;
 		timeouts.ReadTotalTimeoutConstant = timeout_ms.try_into().unwrap_or(u32::MAX);
 		unsafe {
-			SetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &timeouts)
+			SetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 
@@ -187,13 +186,13 @@ impl RawSyncSerialPort {
 
 		let mut timeouts = unsafe { std::mem::zeroed() };
 		unsafe {
-			GetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &mut timeouts)
+			GetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &mut timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		timeouts.WriteTotalTimeoutMultiplier = u32::MAX;
 		timeouts.WriteTotalTimeoutConstant = timeout_ms.try_into().unwrap_or(u32::MAX);
 		unsafe {
-			SetCommTimeouts(HANDLE(self.fd.as_raw_handle() as isize), &timeouts)
+			SetCommTimeouts(HANDLE(self.fd.as_raw_handle()), &timeouts)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 
@@ -217,7 +216,7 @@ impl RawSyncSerialPort {
 
 		match unsafe {
 			ReadFile(
-				HANDLE(self.fd.as_raw_handle() as isize),
+				HANDLE(self.fd.as_raw_handle()),
 				Some(buff),
 				Some(&mut read_bytes),
 				Some(&mut overlapped),
@@ -291,7 +290,7 @@ impl RawSyncSerialPort {
 
 		match unsafe {
 			WriteFile(
-				HANDLE(self.fd.as_raw_handle() as isize),
+				HANDLE(self.fd.as_raw_handle()),
 				Some(buff),
 				Some(&mut written),
 				Some(&mut overlapped),
@@ -346,8 +345,7 @@ impl RawSyncSerialPort {
 	/// If we cannot call `FlushFileBuffers`, or it returns an error.
 	pub fn flush_output(&self) -> IoResult<()> {
 		unsafe {
-			FlushFileBuffers(HANDLE(self.fd.as_raw_handle() as isize))
-				.map_err(|_| IoError::last_os_error())
+			FlushFileBuffers(HANDLE(self.fd.as_raw_handle())).map_err(|_| IoError::last_os_error())
 		}
 	}
 
@@ -366,11 +364,8 @@ impl RawSyncSerialPort {
 		}
 
 		unsafe {
-			PurgeComm(
-				HANDLE(self.fd.as_raw_handle() as isize),
-				PURGE_COMM_FLAGS(flags),
-			)
-			.map_err(|_| IoError::last_os_error())
+			PurgeComm(HANDLE(self.fd.as_raw_handle()), PURGE_COMM_FLAGS(flags))
+				.map_err(|_| IoError::last_os_error())
 		}
 	}
 
@@ -382,7 +377,7 @@ impl RawSyncSerialPort {
 	pub fn set_rts(&self, state: bool) -> IoResult<()> {
 		unsafe {
 			EscapeCommFunction(
-				HANDLE(self.fd.as_raw_handle() as isize),
+				HANDLE(self.fd.as_raw_handle()),
 				if state { SETRTS } else { CLRRTS },
 			)
 			.map_err(|_| IoError::last_os_error())
@@ -406,7 +401,7 @@ impl RawSyncSerialPort {
 	pub fn set_dtr(&self, state: bool) -> IoResult<()> {
 		unsafe {
 			EscapeCommFunction(
-				HANDLE(self.fd.as_raw_handle() as isize),
+				HANDLE(self.fd.as_raw_handle()),
 				if state { SETDTR } else { CLRDTR },
 			)
 			.map_err(|_| IoError::last_os_error())
@@ -488,7 +483,7 @@ impl RawSyncSerialPort {
 	fn read_pin(fd: &File, pin: u32) -> IoResult<bool> {
 		let mut bits: MODEM_STATUS_FLAGS = MODEM_STATUS_FLAGS(0);
 		unsafe {
-			GetCommModemStatus(HANDLE(fd.as_raw_handle() as isize), &mut bits)
+			GetCommModemStatus(HANDLE(fd.as_raw_handle()), &mut bits)
 				.map_err(|_| IoError::last_os_error())?;
 		}
 		Ok(bits.0 & pin != 0)
@@ -499,7 +494,7 @@ impl RawSyncSerialPort {
 			let mut transferred = 0;
 
 			match GetOverlappedResult(
-				HANDLE(file.as_raw_handle() as isize),
+				HANDLE(file.as_raw_handle()),
 				overlapped,
 				&mut transferred,
 				true,
@@ -547,7 +542,7 @@ struct RegKey {
 }
 impl RegKey {
 	fn open(parent: HKEY, subpath: &CStr, rights: REG_SAM_FLAGS) -> IoResult<Self> {
-		let mut key: HKEY = HKEY(std::ptr::null_mut::<std::ffi::c_void>() as isize);
+		let mut key: HKEY = HKEY(std::ptr::null_mut::<std::ffi::c_void>());
 
 		unsafe {
 			RegOpenKeyExA(parent, PCSTR(subpath.as_ptr().cast()), 0, rights, &mut key)
